requirements.txt
torch
transformers
streamlit
langchain
python-docx
